Plugin efficiency, moving some boilerplate plugin logic to core, Equipping items, levelling up gfx, use item fix, implementing ground items, sounds and sounds duplicating fix, global objects

+++ src/main/typescript/elvarg/net/packet/impl/DropItemPacketListener.ts
@@ -1,17 +1,15 @@
 import { PacketExecutor } from "../PacketExecutor";
 import { Packet } from "../Packet";
-// import { Player } from "../../../game/entity/impl/player/Player";
-// import { Item } from "../../../game/model/Item";
-// import { Sound } from "../../../game/Sound";
-// import { Inventory } from "../../../game/model/container/impl/Inventory";
-// import { ItemOnGroundManager } from "../../../game/entity/impl/grounditem/ItemOnGroundManager";
-// import { WildernessArea } from "../../../game/model/areas/impl/WildernessArea";
-// import { Sounds } from "../../../game/Sounds";
-// import { PetHandler } from '../../../game/content/PetHandler'
-// import { PlayerRights } from "../../../game/model/rights/PlayerRights";
+import { Item } from "../../../game/model/Item";
+import { Sound } from "../../../game/Sound";
+import { Inventory } from "../../../game/model/container/impl/Inventory";
+import { ItemOnGroundManager } from "../../../game/entity/impl/grounditem/ItemOnGroundManager";
+import { WildernessArea } from "../../../game/model/areas/impl/WildernessArea";
+import { Sounds } from "../../../game/Sounds";
+import { PetHandler } from "../../../game/content/PetHandler";
+import { PlayerRights } from "../../../game/model/rights/PlayerRights";
 
 export class DropItemPacketListener implements PacketExecutor {
-  // public static destroyItemInterface(player: Player, item: Item) {
   public static destroyItemInterface(player: any, item: any) {
     player.setDestroyItem(item.getId());
     let info: { [key: string]: string }[] = [
@@ -23,27 +21,24 @@ export class DropItemPacketListener implements PacketExecutor {
       { "You cannot get it back if discarded.": "14183" },
       { [item.getDefinition().getName()]: "14184" },
     ];
-    player
-      .getPacketSender()
-      .sendItemOnInterfaces(14171, item.getId(), item.getAmount());
+    player.getPacketSender().sendItemOnInterface(14171, item.getId(), 0, item.getAmount());
     for (let i = 0; i < info.length; i++)
       player.getPacketSender().sendString(info[i][0], parseInt(info[i][1]));
     player.getPacketSender().sendChatboxInterface(14170);
   }
 
-  // public execute(player: Player, packet: Packet) {
   public execute(player: any, packet: Packet) {
     let id = packet.readUnsignedShortA();
-    let interface_id = packet.readUnsignedShort();
+    let interfaceId = packet.readUnsignedShort();
     let itemSlot = packet.readUnsignedShortA();
 
     if (player == null || player.getHitpoints() <= 0) {
       return;
     }
 
-    // if (interface_id != Inventory.INTERFACE_ID) {
-    //     return;
-    // }
+    if (interfaceId != Inventory.INTERFACE_ID) {
+      return;
+    }
 
     if (player.getHitpoints() <= 0) return;
 
@@ -59,9 +54,9 @@ export class DropItemPacketListener implements PacketExecutor {
       return;
     }
 
-    // if (player.getRights() == PlayerRights.DEVELOPER) {
-    //     player.getPacketSender().sendMessage("Drop item: " + item.getId().toString() + ".");
-    // }
+    if (player.getRights() == PlayerRights.DEVELOPER) {
+      player.getPacketSender().sendMessage("Drop item: " + item.getId().toString() + ".");
+    }
 
     player.getPacketSender().sendInterfaceRemoval();
 
@@ -69,20 +64,21 @@ export class DropItemPacketListener implements PacketExecutor {
     player.getSkillManager().stopSkillable();
 
     // Check if we're dropping a pet..
-    // if (PetHandler.drop(player, id, false)) {
-    //     Sounds.sendSound(player, Sound.DROP_ITEM);
-    //     return;
-    // }
+    if (PetHandler.drop(player, id, false)) {
+      Sounds.sendSound(player, Sound.DROP_ITEM);
+      return;
+    }
 
     if (item.getDefinition().isDropable()) {
-      // Items dropped in the Wilderness should go global immediately.
-      // if (player.getArea() instanceof WildernessArea) {
-      //     ItemOnGroundManager.registerGlobal(player, item);
-      // } else {
-      //     ItemOnGroundManager.registers(player, item);
-      // }
-      // player.getInventory().setItem(itemSlot, new Item(-1, 0)).refreshItems();
-      // Sounds.sendSound(player, Sound.DROP_ITEM);
+      const toFloor = new Item(item.getId(), item.getAmount());
+      if (player.getArea() instanceof WildernessArea) {
+        ItemOnGroundManager.registerGlobal(player, toFloor);
+      } else {
+        ItemOnGroundManager.registers(player, toFloor);
+      }
+
+      player.getInventory().setItem(itemSlot, new Item(-1, 0)).refreshItems();
+      Sounds.sendSound(player, Sound.DROP_ITEM);
     } else {
       DropItemPacketListener.destroyItemInterface(player, item);
     }
